# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "OpenApi Documentation",
        "description": "Документация для CDP сервиса",
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "http://127.0.0.1:8000",
            "description": "Основной API"
        }
    ],
    "paths": {
        "/api/dashboard/login": {
            "post": {
                "tags": [
                    "admin"
                ],
                "summary": "Авторизация админа",
                "operationId": "loginAdmin",
                "parameters": [
                    {
                        "name": "login",
                        "in": "path",
                        "description": "Логин администратора",
                        "required": true,
                        "schema": {
                            "type": "string"
                        },
                        "example": "test"
                    },
                    {
                        "name": "password",
                        "in": "path",
                        "description": "Пароль администратора",
                        "required": true,
                        "schema": {
                            "type": "string"
                        },
                        "example": "password"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "array",
                                    "items": {
                                        "properties": {
                                            "success": {
                                                "type": "boolean"
                                            },
                                            "bearerToken": {
                                                "type": "string"
                                            }
                                        },
                                        "type": "object"
                                    }
                                }
                            }
                        }
                    }
                }
            }
        },
...
    },
    "components": {
        "securitySchemes": {
            "Bearer": {
                "type": "apiKey",
                "name": "Authorization",
                "in": "header",
                "scheme": "Bearer"
            }
        }
    }
}